chore: update build and scan stage to CI workflow (#547)
* chore: update build and scan stage to CI workflow

Improves the CI workflow by adding a build and scan stage using Trivy to detect vulnerabilities in the Docker image.
The scan results are uploaded to the GitHub Security tab.
The publish stage now depends on the build and scan stage, ensuring that only images that pass the vulnerability scan are published.

---------

Co-authored-by: ngocdung1004 <nguyenngocdung100420032k3@gmail.com>
Co-authored-by: Nguyễn Ngọc Dũng <96509560+ngocdung1004@users.noreply.github.com>
Co-authored-by: sangtran13 <tranngocsang147877@gmail.com>